requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
matplotlib
seaborn
python-dotenv


